requirements.txt
flask
flask-cors
gunicorn
google-genai
requests
